dom/media/mediasource/test/test_SeekableAfterEndOfStream_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seekable attribute after end of stream</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var updateCount = 0;

runWithMSE(function (ms, v) {
  ms.addEventListener("sourceopen", function () {
    var sb = ms.addSourceBuffer("video/mp4");

    fetchWithXHR("bipbop/bipbop2s.mp4", function (arrayBuffer) {
      sb.appendBuffer(new Uint8Array(arrayBuffer));
      sb.addEventListener("updateend", function () {
        updateCount++;
        /* Ensure that we endOfStream on the first update event only as endOfStream can
           raise more if the duration of the last buffered range and the intial duration
           differ. See bug 1065207 */
        if (updateCount == 1) {
          ms.endOfStream();
        };
      });
    });

    var target = 1.3;

    v.addEventListener("loadedmetadata", function () {
      ok(v.seekable.length, "Resource is seekable");
      ok(v.seekable.length &&
          target >= v.seekable.start(0) &&
          target < v.seekable.end(0), "Target is within seekable range");
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
